messing about, not going to hot
So weve implemented some form of color changing. But, it seems like having a
single background color is the only viable option. What we CAN do is create a
"highlight" color, which could be used to draw the rectangles used for
highlighting. All in all a better and different approach is needed to handle
colors and highlighting.